apps/ai-agent/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
google-generativeai==0.8.5
requests==2.32.5
pydantic==2.11.9